firebaseData.js creado,firebaseAuth y firebaseInit modificados para correr funcion SaveRegistryData
