+++ src/assets/js/firebase_auth.js
@@ -1,20 +1,56 @@
+import { validationFormSignLogin } from '../controller/validation.js'
+import { SaveRegistryData } from './firebase_data.js'
+//HU1 registro con cuenta Google
+// Function loginGoogle
+export const loginGoogle = () => {
+  console.log("click en LOGINGOOGLE hacemos correr su funcion? sii!! ");
+  // 1. Crea una instancia del objeto del proveedor de Google:
+  var provider = new firebase.auth.GoogleAuthProvider();
+  // 2.3.4 opcionales
+  // 5. Autenticar con Firebase a través del objeto del proveedor de Google.
+  // Para ofrecer acceso con una ventana emergente, invoca signInWithPopup:
+  firebase.auth().signInWithPopup(provider)
+    .then(function (result) {
+      // This gives you a Google Access Token. You can use it to access the Google API.
+      var token = result.credential.accessToken;
+      // The signed-in user info.
+      var user = result.user;
+      // ...
+      
+      SaveRegistryData();
+    })
+    .then(function changeMuro() {
+      window.location.hash = "#/muro";
+    })
+    .catch(function (error) {
+      // Handle Errors here.
+      var errorCode = error.code;
+      var errorMessage = error.message;
+      // The email of the user's account used.
+      var email = error.email;
+      // The firebase.auth.AuthCredential type that was used.
+      var credential = error.credential;
+      // ...
+    });
+};
+
+//HU2 registro con Correo 
 export const createUser = (user, age, email, password) => {
   console.log(user);
   console.log(age);
   console.log(email);
   console.log(password);
 
-  firebase
-    .auth()
-    .createUserWithEmailAndPassword(email, password)
-
+  firebase.auth().createUserWithEmailAndPassword(email, password)
+    
     .then(function () {
       console.log("se creo usuario en firebase");
+      SaveRegistryData();
       verificationEmail();
       window.location.hash = "#/login";
     })
 
-    
+
     .catch(function (error) {
       // Handle Errors here.
       var errorCode = error.code;
@@ -22,31 +58,7 @@ export const createUser = (user, age, email, password) => {
       // ...
     });
 
- //inicializar con Cloud firestore   
-let db = firebase.firestore();
 
-//agregando a la colletion
-
-function guardar (){
-  let user = document.getElementById('user').value;
-  let age = document.getElementById('age').value;
-  let email = document.getElementById('email').value;
-  let password = document.getElementById('password').value;
-
-  db.collection("users").add({
-    user: user,
-    age: age,
-    email: email,
-    password: password
-})
-.then(function(docRef) {
-    console.log("Document written with ID: ", docRef.id);
-})
-.catch(function(error) {
-    console.error("Error adding document: ", error);
-});
-
-}
   // let db = firebase.firestore();
   // firebase.auth().createUserWithEmailAndPassword(email, password)
   // .then(function(){
@@ -67,49 +79,21 @@ function guardar (){
   // });
 };
 
-// Function loginGoogle
-export const loginGoogle = () => {
-  console.log("click en LOGINGOOGLE hacemos correr su funcion? sii!! ");
-  // 1. Crea una instancia del objeto del proveedor de Google:
-  var provider = new firebase.auth.GoogleAuthProvider();
-  // 2.3.4 opcionales
-  // 5. Autenticar con Firebase a través del objeto del proveedor de Google.
-  // Para ofrecer acceso con una ventana emergente, invoca signInWithPopup:
-  firebase
-    .auth()
-    .signInWithPopup(provider)
-    .then(function (result) {
-      // This gives you a Google Access Token. You can use it to access the Google API.
-      var token = result.credential.accessToken;
-      // The signed-in user info.
-      var user = result.user;
-      // ...
-    })
-    .then(function changeMuro() {
-      window.location.hash = "#/muro";
-    })
-    .catch(function (error) {
-      // Handle Errors here.
-      var errorCode = error.code;
-      var errorMessage = error.message;
-      // The email of the user's account used.
-      var email = error.email;
-      // The firebase.auth.AuthCredential type that was used.
-      var credential = error.credential;
-      // ...
-    });
-};
 
+
+
+//HU3 Inicio de Sesion 
 export const signLogin = (email, password) => {
   console.log("antes de realizar signLogin ");
   // let user = firebase.auth().currentUser;
   // var emailVerifiedReturn = user.emailVerified;
 
   firebase.auth().signInWithEmailAndPassword(email, password)
-  
-    
+
+
     .then(
       function () {
+        validationFormSignLogin();
         observer();
         // window.location.hash = "#/muro";
         //si user.verified es true entonces se va al muro
@@ -128,6 +112,7 @@ export const signLogin = (email, password) => {
     });
 };
 
+// Observer
 export const observer = () => {
   firebase.auth().onAuthStateChanged(function (user) {
     if (user) {
@@ -156,16 +141,18 @@ export const observer = () => {
       // window.location.hash = "#/home";
       // User is signed out.
       // ...
-      return false; 
+      return false;
     }
   });
   //email-password.html
 };
 
+
+//Funcion Aparece
 function aparece(user) {
   var user = user;
   if (user.emailVerified) {
-     window.location.hash = "#/muro";
+    window.location.hash = "#/muro";
 
   }
 
@@ -173,10 +160,11 @@ function aparece(user) {
     console.log("el correo no ha sido verificado");
     swal.fire("puedes verificar tu correo en la bandeja de entrada");
     // window.location.hash = "#/login";
-    
+
   }
 }
 
+//Función VerificacionCorreo
 export const verificationEmail = () => {
   // para enviar un mensaje de direccion a un usuario ...
   var user = firebase.auth().currentUser;
@@ -194,6 +182,7 @@ export const verificationEmail = () => {
     });
 };
 
+//Función Cerrar Sesión
 export const signOutRedSocial = () => {
   firebase
     .auth()
